Assignment 42 partially completed

+++ app/controllers/comments_controller.rb
@@ -3,29 +3,56 @@ class CommentsController < ApplicationController
   before_action :authorize_user, only: [:destroy]
 
   def create
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.new(comment_params)
-    comment.user = current_user
+    if params[:topic_id]
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.new(comment_params)
+      comment.user = current_user
 
-    if comment.save
-      flash[:notice] = "Comment saved successfully."
-      redirect_to [@post.topic, @post]
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to @topic
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to @topic
+      end
     else
-      flash[:alert] = "Comment failed to save."
-      redirect_to [@post.topic, @post]
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@post.topic, @post]
+      end
     end
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.find(params[:id])
+    if params[:topic_id]
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.find(params[:id])
 
-    if comment.destroy
-      flash[:notice] = "Comment was deleted."
-      redirect_to [@post.topic, @post]
+      if comment.destroy
+        flash[:notic] = "Comment was deleted."
+        redirect_to @topic
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to @topic
+      end
     else
-      flash[:alert] = "Comment couldn't be deleted. Try again."
-      redirect_to [@post.topic, @post]
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.find(params[:id])
+
+      if comment.destroy
+        flash[:notice] = "Comment was deleted."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@post.topic, @post]
+      end
     end
   end
 
@@ -39,7 +66,11 @@ class CommentsController < ApplicationController
     comment = Comment.find(params[:id])
     unless current_user == comment.user || current_user.admin?
       flash[:alert] = "You do not have permission to delete a comment."
-      redirect_to [comment.post.topic, comment.post]
+      if params[:topic_id]
+        redirect_to comment.topic
+      else
+        redirect_to [comment.post.topic, comment.post]
+      end
     end
   end
 end
